🏗️ separate layers for different responsibilities
