+++ tests/test_main.py
@@ -1,9 +1,10 @@
 from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
+from pyrb.controller import app
 from pyrb.exceptions import InsufficientFundsException
-from pyrb.main import RebalanceContext, app
-from pyrb.order import Order, OrderType
+from pyrb.repositories.context import RebalanceContext
+from pyrb.repositories.order_manager import Order, OrderType
 
 
 def test_sut_rebalances(fake_rebalance_context: RebalanceContext, mocker: MockerFixture) -> None:
@@ -12,7 +13,7 @@ def test_sut_rebalances(fake_rebalance_context: RebalanceContext, mocker: Mocker
     # given
     runner = CliRunner()
 
-    mocker.patch("pyrb.main._create_rebalance_context", return_value=fake_rebalance_context)
+    mocker.patch("pyrb.controller.create_rebalance_context", return_value=fake_rebalance_context)
 
     spy = mocker.spy(fake_rebalance_context.order_manager, "place_order")
 
@@ -50,7 +51,7 @@ def test_sut_rebalances_with_only_buy_orders(
     # given
     runner = CliRunner()
 
-    mocker.patch("pyrb.main._create_rebalance_context", return_value=fake_rebalance_context)
+    mocker.patch("pyrb.controller.create_rebalance_context", return_value=fake_rebalance_context)
 
     # when
     result = runner.invoke(app, ["rebalance", "--investment-amount", "20000"])
@@ -66,7 +67,7 @@ def test_sut_rebalances_with_only_sell_order(
     # given
     runner = CliRunner()
 
-    mocker.patch("pyrb.main._create_rebalance_context", return_value=fake_rebalance_context)
+    mocker.patch("pyrb.controller.create_rebalance_context", return_value=fake_rebalance_context)
 
     # when
     result = runner.invoke(app, ["rebalance", "--investment-amount", "1000"])
@@ -82,7 +83,7 @@ def test_sut_stops_rebalancing_with_disallowance_from_user(
     # given
     runner = CliRunner()
 
-    mocker.patch("pyrb.main._create_rebalance_context", return_value=fake_rebalance_context)
+    mocker.patch("pyrb.controller.create_rebalance_context", return_value=fake_rebalance_context)
 
     # when
     result = runner.invoke(app, ["rebalance", "--investment-amount", "1000"], input="n\n")
@@ -98,7 +99,7 @@ def test_sut_stops_rebalancing_with_insufficient_funds(
     # given
     runner = CliRunner()
 
-    mocker.patch("pyrb.main._create_rebalance_context", return_value=fake_rebalance_context)
+    mocker.patch("pyrb.controller.create_rebalance_context", return_value=fake_rebalance_context)
 
     # when
     result = runner.invoke(app, ["rebalance", "--investment-amount", "999999999"], input="y\n")
